+++ src/Kiota.Builder/CodeDOM/CodeBlock.cs
@@ -30,7 +30,7 @@ public class CodeBlock<TBlockDeclaration, TBlockEnd> : CodeElement, IBlock where
             return InnerChildElements.Values;
         return new CodeElement[] { StartBlock, EndBlock }.Union(InnerChildElements.Values);
     }
-    public void RenameChildElement(string oldName, string newName)
+    public virtual void RenameChildElement(string oldName, string newName)
     {
         if (InnerChildElements.TryRemove(oldName, out var element))
         {
@@ -44,7 +44,7 @@ public class CodeBlock<TBlockDeclaration, TBlockEnd> : CodeElement, IBlock where
         if (elements == null) return;
         RemoveChildElementByName(elements.Select(static x => x.Name).ToArray());
     }
-    public void RemoveChildElementByName(params string[] names)
+    public virtual void RemoveChildElementByName(params string[] names)
     {
         if (names == null) return;