+++ tests/Kiota.Builder.Tests/KiotaBuilderTests.cs
@@ -6088,4 +6088,144 @@ components:
         Assert.Equal("sub1", codeModel.FindChildByName<CodeClass>("sub1").Documentation.Description);
         Assert.Equal("sub2", codeModel.FindChildByName<CodeClass>("sub2").Documentation.Description);
     }
+    [Fact]
+    public async Task CleanupSymbolNameDoesNotCauseNameConflicts()
+    {
+        var tempFilePath = Path.Combine(Path.GetTempPath(), Path.GetTempFileName());
+        await using var fs = await GetDocumentStream(@"openapi: 3.0.1
+info:
+  title: Example
+  description: Example
+  version: 1.0.1
+servers:
+  - url: https://example.org
+paths:
+  /directoryObject:
+    get:
+      responses:
+        '200':
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/entity'
+components:
+  schemas:
+    entity:
+      title: entity
+      type: object
+      required: ['type', '@type']
+      properties:
+        type:
+          type: string
+        '@type':
+          type: integer");
+        var mockLogger = new Mock<ILogger<KiotaBuilder>>();
+        var builder = new KiotaBuilder(mockLogger.Object, new GenerationConfiguration { ClientClassName = "Graph", OpenAPIFilePath = tempFilePath, IncludeAdditionalData = false }, _httpClient);
+        var document = await builder.CreateOpenApiDocumentAsync(fs);
+        var node = builder.CreateUriSpace(document);
+        var codeModel = builder.CreateSourceModel(node);
+        var resultClass = codeModel.FindChildByName<CodeClass>("Entity");
+        Assert.NotNull(resultClass);
+        Assert.Equal(2, resultClass.Properties.Select(static x => x.Name).Distinct(StringComparer.OrdinalIgnoreCase).Count());
+    }
+    [Fact]
+    public async Task CleanupSymbolNameDoesNotCauseNameConflictsWithSuperType()
+    {
+        var tempFilePath = Path.Combine(Path.GetTempPath(), Path.GetTempFileName());
+        await using var fs = await GetDocumentStream(@"openapi: 3.0.1
+info:
+  title: Example
+  description: Example
+  version: 1.0.1
+servers:
+  - url: https://example.org
+paths:
+  /directoryObject:
+    get:
+      responses:
+        '200':
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/subtype'
+components:
+  schemas:
+    entity:
+      title: entity
+      type: object
+      required: ['@type']
+      properties:
+        '@type':
+          type: integer
+      discriminator:
+        propertyName: '@type'
+        mapping:
+          'subtype': '#/components/schemas/subtype'
+    subtype:
+      allOf:
+        - $ref: '#/components/schemas/entity'
+        - title: subtype
+          type: object
+          required: ['type', '@type']
+          properties:
+            'type':
+              type: string");
+        var mockLogger = new Mock<ILogger<KiotaBuilder>>();
+        var builder = new KiotaBuilder(mockLogger.Object, new GenerationConfiguration { ClientClassName = "Graph", OpenAPIFilePath = tempFilePath, IncludeAdditionalData = false }, _httpClient);
+        var document = await builder.CreateOpenApiDocumentAsync(fs);
+        var node = builder.CreateUriSpace(document);
+        var codeModel = builder.CreateSourceModel(node);
+        var entityClass = codeModel.FindChildByName<CodeClass>("Entity");
+        Assert.NotNull(entityClass);
+        var atType = entityClass.FindChildByName<CodeProperty>("Type");
+        Assert.Equal("@type", atType.WireName);
+        var subtypeClass = codeModel.FindChildByName<CodeClass>("Subtype");
+        Assert.NotNull(subtypeClass);
+        var type = subtypeClass.FindChildByName<CodeProperty>("SubtypeType");
+        Assert.Equal("type", type.WireName);
+        Assert.Equal("subtypeType", type.Name);
+    }
+    [Fact]
+    public async Task CleanupSymbolNameDoesNotCauseNameConflictsInQueryParameters()
+    {
+        var tempFilePath = Path.Combine(Path.GetTempPath(), Path.GetTempFileName());
+        await using var fs = await GetDocumentStream(@"openapi: 3.0.1
+info:
+  title: Example
+  description: Example
+  version: 1.0.1
+servers:
+  - url: https://example.org
+paths:
+  /directoryObject:
+    get:
+      parameters:
+        - name: $select
+          in: query
+          schema:
+            type: string
+        - name: select
+          in: query
+          schema:
+            type: int64
+      responses:
+        '200':
+          content:
+            application/json:
+              schema:
+                type: string");
+        var mockLogger = new Mock<ILogger<KiotaBuilder>>();
+        var builder = new KiotaBuilder(mockLogger.Object, new GenerationConfiguration { ClientClassName = "Graph", OpenAPIFilePath = tempFilePath, IncludeAdditionalData = false }, _httpClient);
+        var document = await builder.CreateOpenApiDocumentAsync(fs);
+        var node = builder.CreateUriSpace(document);
+        var codeModel = builder.CreateSourceModel(node);
+        var parametersClass = codeModel.FindChildByName<CodeClass>("directoryObjectRequestBuilderGetQueryParameters");
+        Assert.NotNull(parametersClass);
+        var dollarSelect = parametersClass.FindChildByName<CodeProperty>("Select");
+        Assert.Equal("%24select", dollarSelect.WireName);
+        Assert.Equal("string", dollarSelect.Type.Name);
+        var select = parametersClass.FindChildByName<CodeProperty>("select0");
+        Assert.Equal("select", select.WireName);
+        Assert.Equal("int64", select.Type.Name);
+    }
 }